requirements.txt
chromadb
flask
pandas
openpyxl
pydantic
uuid
python-dotenv
